refactor(cli): extract snapshot actions
